rename registry.h and tape.h to work around macOS case insensitivity
WARNING: I don't have the dependencies installed to test the qt build. Sorry!
