+++ source/linux/duplicates/Registry.cpp
@@ -1,8 +1,5 @@
 #include "StdAfx.h"
 #include "Registry.h"
-#include "Common.h"
-#include "Card.h"
-#include "CopyProtectionDongles.h"
