+++ source/linux/registryclass.h
